+++ aegis/flo-aegis.h
@@ -16,25 +16,25 @@ extern "C"{
 #endif
 
 int crypto_aead_encrypt_opt(
-	unsigned char *c,unsigned long long *clen,
-	unsigned char *m,unsigned long long mlen,
-	unsigned char *ad,unsigned long long adlen,
-	unsigned char *npub,
-	unsigned char *k);
+		unsigned char *c, uint64_t *clen,
+		unsigned char *m, uint64_t mlen,
+		unsigned char *ad, uint64_t adlen,
+		unsigned char *npub,
+		unsigned char *k);
 
 int crypto_aead_encrypt(
-		unsigned char *c,unsigned long long *clen,
-		const unsigned char *m,unsigned long long mlen,
-		const unsigned char *ad,unsigned long long adlen,
+		unsigned char *c, uint64_t *clen,
+		const unsigned char *m, uint64_t mlen,
+		const unsigned char *ad, uint64_t adlen,
 		const unsigned char *npub,
 		const unsigned char *k
 );
 
 int crypto_aead_decrypt(
-		unsigned char *m,unsigned long long *mlen,
+		unsigned char *m, uint64_t *mlen,
 		unsigned char *nsec,
-		const unsigned char *c,unsigned long long clen,
-		const unsigned char *ad,unsigned long long adlen,
+		const unsigned char *c, uint64_t clen,
+		const unsigned char *ad, uint64_t adlen,
 		const unsigned char *npub,
 		const unsigned char *k
 );
